Added parameters validator, improved file structure

+++ public/routes/WebRoutes.php
@@ -19,8 +19,10 @@ switch($uri) {
         $controller = new FoodController();
         if($_SERVER["REQUEST_METHOD"] === 'POST') {
             if(isset($_POST['name']) && !empty($_POST['name'])) {
-                $name = trim($_POST['name']);
-                $controller->addFood($name);
+                if(is_string($_POST["name"]) && (strlen($_POST["name"]) > 0 && strlen($_POST["name"]) < 50)) {
+                    $name = trim($_POST['name']);
+                    $controller->addFood($name);
+                }
             }
         } else {
             echo $controller->create();
@@ -30,6 +32,10 @@ switch($uri) {
         $controller = new FoodController();
         if($_SERVER["REQUEST_METHOD"] === 'POST') {
             if(isset($_POST['id']) && !empty($_POST['id'])) {
+                if(!is_numeric($_POST['id']) || $_POST['id'] <= 0) {
+                    header("Location: /food");
+                    die();
+                }
                 $id = $_POST['id'];
                 $controller->delete($id);
                 break;
@@ -45,10 +51,6 @@ switch($uri) {
         $controller = $container->get(UserController::class);
 
         if($_SERVER["REQUEST_METHOD"] === 'POST') {
-            if(isset($_POST['name']) && !empty($_POST['name']) &&
-               isset($_POST['lastName']) && !empty($_POST['lastName']) &&
-               isset($_POST['email']) && !empty($_POST['email'])) {
-
                 $name = trim($_POST['name']);
                 $lastName = trim($_POST['lastName']);
                 $email = trim($_POST['email']);
@@ -59,7 +61,6 @@ switch($uri) {
                 $controller->addUser(
                     $data
                 );
-            }
             break;
         }
 
@@ -70,10 +71,6 @@ switch($uri) {
         $controller = $container->get(UserController::class);
 
         if($_SERVER["REQUEST_METHOD"] === 'POST') {
-            if(isset($_POST['id']) && !empty($_POST['id']) &&
-               isset($_POST['name']) && !empty($_POST['name']) &&
-               isset($_POST['lastName']) && !empty($_POST['lastName']) &&
-               isset($_POST['email']) && !empty($_POST['email'])) {
 
                 $id = $_POST['id'];
                 $name = trim($_POST['name']);
@@ -87,11 +84,10 @@ switch($uri) {
                     id: $id,
                     data: $data
                 );
-            }
             break;
         }
         
-        if(!isset($_GET['id']) || empty($_GET['id'])) {
+        if(!isset($_GET['id']) || empty($_GET['id']) || !is_numeric($_GET['id']) || $_GET['id'] <= 0) {
             header("Location: /users");
             die();
         }
@@ -111,10 +107,10 @@ switch($uri) {
         header("Location: /users");
         break;
     case "/report/create":
-        require_once \BASE_PATH . "/public/reportRequest.php";
+        require_once \BASE_PATH . "/src/views/reports/reportRequest.php";
         break;
     case "/report/download":
-        require \BASE_PATH . "/public/downloadReport.php";
+        require \BASE_PATH . "/src/views/reports/downloadReport.php";
         break;
     default:
         echo "404 Not Found";
