+++ src/Controllers/UserController.php
@@ -5,8 +5,9 @@ use Doctrine\ORM\EntityManager;
 use App\Entity\User;
 use App\config\EntityManagerFactory;
 use App\Repositories\UserRepository;
+use App\Validator\Validator;
 
-class UserController
+class UserController extends BaseController
 {
     public function __construct(private UserRepository $userRepository,private EntityManager $entityManager)
     {
@@ -16,20 +17,35 @@ class UserController
     {
         $users = $this->userRepository->findAll();
 
-        // var_dump($users);
-
-        require_once \BASE_PATH . '/public/user_views/user.php';
+        $this->view("users","index",["users" => $users]);
     }
 
     public function create()
     {
         $foods = $this->entityManager->getRepository(\App\Entity\Food::class)->findAll();
-        require_once \BASE_PATH . '/public/user_views/addUser.php';
+        $this->view("users","create",["foods" => $foods]);
     }
 
     public function addUser(array $data)
     {
-        $user = $this->userRepository->storeUser($data);
+        $validator = new Validator();
+        $rules = [
+            'name' => 'required|string|min:2|max:50',
+            'lastName' => 'required|string|min:2|max:50',
+            'email' => 'required|email',
+            'foodIds' => 'array|numeric_array'
+        ];
+
+        if($validator->validate($data, $rules)){
+            $_SESSION['errors'] = $validator->getErrors();
+            header("Location: /user/create");
+            die();
+        }
+
+        $user = $this->userRepository->storeUser($data["name"],
+            $data["lastName"],
+            $data["email"],
+            $data["foodIds"] ?? []);
 
         header("Location: /users");
         die();
@@ -37,6 +53,17 @@ class UserController
 
     public function delete(int $id)
     {
+        $validator = new Validator();
+        $rules = [
+            'id' => 'required|numeric'
+        ];
+
+         if($validator->validate(["id" => $id], $rules)){
+            $_SESSION['errors'] = $validator->getErrors();
+            header("Location: /users");
+            die();
+        }
+
         $this->userRepository->deleteUser($id);
 
         header("Location: /users");
@@ -56,11 +83,25 @@ class UserController
         $userFoodIds = array_map(fn($food) => $food->getId(), $userFoods);
 
         $foods = $this->entityManager->getRepository(\App\Entity\Food::class)->findAll();
-        require_once \BASE_PATH . '/public/user_views/editUser.php';
+        $this->view("users","edit",["user" => $user, "userFoodIds" => $userFoodIds,"foods" => $foods]);
     }
 
     public function update(int $id,array $data):void
     {
+        $validator = new Validator();
+        $rules = [
+            'name' => 'required|string|min:2|max:50',
+            'lastName' => 'required|string|min:2|max:50',
+            'email' => 'required|email',
+            'foodIds' => 'numeric_array'
+        ];
+
+        if($validator->validate($data, $rules)){
+            $_SESSION['errors'] = $validator->getErrors();
+            header("Location: /user/update?id=$id");
+            die();
+        }
+
         $this->userRepository->updateUser($id, $data);
 
         header("Location: /users");
